requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
openai>=1.37.0
reportlab==4.2.2
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9 
typing-extensions>=4.12.2